DRTII-696 Add titles to remaining pages
