+++ react/src/App.tsx
@@ -1,66 +1,66 @@
-import React, {useEffect} from 'react';
-import './App.css';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import { useTheme } from '@mui/material/styles';
-import {Home} from './components/Home';
-import Alerts from './components/alerts/Alerts';
-import AccessRequests from './components/accessrequests/AccessRequests';
-import UsersList from './components/users/UsersList';
-import {Route, Routes} from "react-router-dom";
-import Loading from "./components/Loading";
-import Navigation from "./components/Navigation";
-import {useConfig} from "./store/config";
-import {Container} from "@mui/material";
-import {styled} from "@mui/material/styles";
-import {RegionPage} from "./components/RegionPage";
-import axios from "axios";
-import ApiClient from "./services/ApiClient";
-import {HealthCheckEditor} from "./components/healthcheckpauseseditor/HealthCheckPausesEditor";
-import {useUser} from "./store/user";
-import {DropInSessionsList} from "./components/dropins/DropInSessionsList";
-import {AddOrEditDropInSession} from "./components/dropins/AddOrEditDropInSession";
-import {DropInSessionRegistrations} from "./components/dropins/DropInSessionRegistrations";
-import {SnackbarProvider} from 'notistack';
-import Link from "@mui/material/Link";
-import {FeatureGuideList} from "./components/featureguides/FeatureGuideList";
-import {FeatureGuideAddOrEdit} from "./components/featureguides/FeatureGuideAddOrEdit";
-import {FeedbackForms} from "./components/feedback/FeedbackForms";
-import {FeedbackList} from "./components/feedback/FeedbackList";
-import DownloadManager from './components/downloadmanager/DownloadManager';
-import {ExportConfig} from "./components/ExportConfig";
-import {HealthChecks} from "./components/healthchecks/HealthChecks";
-import RegionalDashboard from './components/regionalpressure/RegionalDashboard';
-import RegionalDashboardDetail from './components/regionalpressure/RegionalDashboardDetail';
+import React, {useEffect} from 'react'
+import './App.css'
+import useMediaQuery from '@mui/material/useMediaQuery'
+import { useTheme } from '@mui/material/styles'
+import {Home} from './components/Home'
+import Alerts from './components/alerts/Alerts'
+import AccessRequests from './components/accessrequests/AccessRequests'
+import UsersList from './components/users/UsersList'
+import {Route, Routes} from "react-router-dom"
+import Loading from "./components/Loading"
+import Navigation from "./components/Navigation"
+import {useConfig} from "./store/config"
+import {Container} from "@mui/material"
+import {styled} from "@mui/material/styles"
+import {RegionPage} from "./components/RegionPage"
+import axios from "axios"
+import ApiClient from "./services/ApiClient"
+import {HealthCheckEditor} from "./components/healthcheckpauseseditor/HealthCheckPausesEditor"
+import {useUser} from "./store/user"
+import {DropInSessionsList} from "./components/dropins/DropInSessionsList"
+import {AddOrEditDropInSession} from "./components/dropins/AddOrEditDropInSession"
+import {DropInSessionRegistrations} from "./components/dropins/DropInSessionRegistrations"
+import {SnackbarProvider} from 'notistack'
+import Link from "@mui/material/Link"
+import {FeatureGuideList} from "./components/featureguides/FeatureGuideList"
+import {FeatureGuideAddOrEdit} from "./components/featureguides/FeatureGuideAddOrEdit"
+import {FeedbackForms} from "./components/feedback/FeedbackForms"
+import {FeedbackList} from "./components/feedback/FeedbackList"
+import DownloadManager from './components/downloadmanager/DownloadManager'
+import {ExportConfig} from "./components/ExportConfig"
+import {HealthChecks} from "./components/healthchecks/HealthChecks"
+import RegionalDashboard from './components/regionalpressure/RegionalDashboard'
+import NationalDashboard from "./components/regionalpressure/NationalDashboard"
 
 const StyledDiv = styled('div')(() => ({
   textAlign: 'center',
-}));
+}))
 
 const StyledContainer = styled(Container)(() => ({
   textAlign: 'left',
   minHeight: 500,
   display: 'inline-block',
-}));
+}))
 
 export const App = () => {
   const {user} = useUser()
   const {config} = useConfig()
 
-  const currentLocation = window.document.location;
+  const currentLocation = window.document.location
   const logoutLink = "/oauth2/sign_out?redirect=" + currentLocation.toString()
-  const theme = useTheme();
-  const is_mobile = useMediaQuery(theme.breakpoints.down('md'));
+  const theme = useTheme()
+  const is_mobile = useMediaQuery(theme.breakpoints.down('md'))
 
   useEffect(() => {
     const trackUser = async () =>
       axios
         .get(ApiClient.userTrackingEndPoint)
         .catch(reason => {
-          console.log('Unable to user tracking' + reason);
+          console.log('Unable to user tracking' + reason)
         })
 
-    trackUser();
-  }, []);
+    trackUser()
+  }, [])
 
   return (user.kind === "SignedInUser" && config.kind === "LoadedConfig") ?
     <StyledDiv>
@@ -101,8 +101,8 @@ export const App = () => {
           <Route path={"/access-requests"} element={<AccessRequests/>}/>
           <Route path={"/users"} element={<UsersList/>}/>
           <Route path={"/download"} element={<DownloadManager config={config.values} user={user.profile} />} />
-          <Route path={"/national-pressure"} element={<RegionalDashboard config={config.values} user={user.profile} />} />
-          <Route path={"/national-pressure/:region"} element={<RegionalDashboardDetail config={config.values} user={user.profile} />} />
+          <Route path={"/national-pressure"} element={<NationalDashboard config={config.values} user={user.profile} />} />
+          <Route path={"/national-pressure/:region"} element={<RegionalDashboard config={config.values} user={user.profile} />} />
           <Route path={"/alerts"} element={<Alerts regions={config.values.portsByRegion} user={user.profile}/>}/>
           <Route path={"/region/:regionName"} element={<RegionPage user={user.profile} config={config.values}/>}/>
           <Route path={"/feature-guides"}>
