+++ react/src/utils/StringUtils.ts
@@ -1,5 +1,5 @@
 export const StringUtils = {
-    ucFirst: (str: string) => {
-        return str.charAt(0).toUpperCase() + str.slice(1)
-    }
+  ucFirst: (str: string) => {
+    return str.charAt(0).toUpperCase() + str.slice(1)
+  }
 }
